+++ frontend/src/features/section/index.ts
@@ -1,4 +1,3 @@
-export type { SectionFormRequest } from './components/SectionForm'
 export * from './components/CreateSectionButton'
 export * from './components/UpdateSectionButton'
 export * from './components/DraggableSections'
